Added the ability to log the tick lines to a file. Refactored the apex_log_monitor application
