+++ cmd/apex_batch_get/apex_load_previous_data.go
@@ -13,7 +13,6 @@ import (
 
 
 func main() {
-
 	flags := common_flags.GetFlags()
 	configureLogs(flags.LogLevel)
